Create JSON file to put students data by themselves.
To automate testing toy problem and sending the result to spread sheet, try to do new way that is creating a place to insert students data by themselves, and use it for testing and recording to spreadsheet.
In student.json, students insert the problem number that they want to test by themseleves, the mocha get the number of test, and find out which test file it should run.

To do: Check the result of test, if test passed, send students information with test result to the spreadsheet.